+++ include/micro-os-plus/architecture-aarch64/defines.h
@@ -10,8 +10,8 @@
  * be obtained from https://opensource.org/licenses/MIT/.
  */
 
-#ifndef MICRO_OS_PLUS_ARCHITECTURE_CORTEXA_DEFINES_H_
-#define MICRO_OS_PLUS_ARCHITECTURE_CORTEXA_DEFINES_H_
+#ifndef MICRO_OS_PLUS_ARCHITECTURE_AARCH64_DEFINES_H_
+#define MICRO_OS_PLUS_ARCHITECTURE_AARCH64_DEFINES_H_
 
 // ----------------------------------------------------------------------------
 
@@ -20,6 +20,6 @@
 
 // ----------------------------------------------------------------------------
 
-#endif // MICRO_OS_PLUS_ARCHITECTURE_CORTEXA_DEFINES_H_
+#endif // MICRO_OS_PLUS_ARCHITECTURE_AARCH64_DEFINES_H_
 
 // ----------------------------------------------------------------------------