+++ include/micro-os-plus/architecture-aarch64/instructions-inlines.h
@@ -10,15 +10,15 @@
  * be obtained from https://opensource.org/licenses/MIT/.
  */
 
-#ifndef MICRO_OS_PLUS_ARCHITECTURE_CORTEXA_INSTRUCTIONS_INLINES_H_
-#define MICRO_OS_PLUS_ARCHITECTURE_CORTEXA_INSTRUCTIONS_INLINES_H_
+#ifndef MICRO_OS_PLUS_ARCHITECTURE_AARCH64_INSTRUCTIONS_INLINES_H_
+#define MICRO_OS_PLUS_ARCHITECTURE_AARCH64_INSTRUCTIONS_INLINES_H_
 
 // ----------------------------------------------------------------------------
 
 #include <stdint.h>
 
 // ----------------------------------------------------------------------------
-// Inline implementations for the Cortex-A architecture instructions.
+// Inline implementations for the AArch64 architecture instructions.
 
 #if defined(__cplusplus)
 extern "C"
@@ -28,7 +28,7 @@ extern "C"
   // --------------------------------------------------------------------------
 
   static inline __attribute__ ((always_inline)) void
-  cortexa_architecture_nop (void)
+  aarch64_architecture_nop (void)
   {
     __asm__ volatile(
 
@@ -41,9 +41,8 @@ extern "C"
   }
 
   static inline __attribute__ ((always_inline)) void
-  cortexa_architecture_bkpt (void)
+  aarch64_architecture_bkpt (void)
   {
-#if defined(__ARM_ARCH_8A)
     __asm__ volatile(
 
         // See 'aarch64-tdep.c' in GDB source, 'aarch64_default_breakpoint'
@@ -54,22 +53,10 @@ extern "C"
         : /* Inputs */
         : /* Clobbers */
     );
-#elif defined(__ARM_ARCH_7A__)
-  __asm__ volatile(
-
-      " bkpt 0 "
-
-      : /* Outputs */
-      : /* Inputs */
-      : /* Clobbers */
-  );
-#else
-#error "Unsupported architecture."
-#endif
   }
 
   static inline __attribute__ ((always_inline)) void
-  cortexa_architecture_wfi (void)
+  aarch64_architecture_wfi (void)
   {
     __asm__ volatile(
 
@@ -84,7 +71,7 @@ extern "C"
   static inline __attribute__ ((always_inline)) void
   micro_os_plus_architecture_nop (void)
   {
-    cortexa_architecture_nop ();
+    aarch64_architecture_nop ();
   }
 
   /**
@@ -93,7 +80,7 @@ extern "C"
   static inline __attribute__ ((always_inline)) void
   micro_os_plus_architecture_brk (void)
   {
-    cortexa_architecture_bkpt ();
+    aarch64_architecture_bkpt ();
   }
 
   /**
@@ -102,7 +89,7 @@ extern "C"
   static inline __attribute__ ((always_inline)) void
   micro_os_plus_architecture_wfi (void)
   {
-    cortexa_architecture_wfi ();
+    aarch64_architecture_wfi ();
   }
 
   // --------------------------------------------------------------------------
@@ -115,30 +102,30 @@ extern "C"
 
 #if defined(__cplusplus)
 
-namespace cortexa::architecture
+namespace aarch64::architecture
 {
   // --------------------------------------------------------------------------
 
   inline __attribute__ ((always_inline)) void
   nop (void)
   {
-    cortexa_architecture_nop ();
+    aarch64_architecture_nop ();
   }
 
   inline __attribute__ ((always_inline)) void
   bkpt (void)
   {
-    cortexa_architecture_bkpt ();
+    aarch64_architecture_bkpt ();
   }
 
   inline __attribute__ ((always_inline)) void
   wfi (void)
   {
-    cortexa_architecture_wfi ();
+    aarch64_architecture_wfi ();
   }
 
   // --------------------------------------------------------------------------
-} // namespace cortexa::architecture
+} // namespace aarch64::architecture
 
 namespace micro_os_plus::architecture
 {
@@ -147,19 +134,19 @@ namespace micro_os_plus::architecture
   inline __attribute__ ((always_inline)) void
   nop (void)
   {
-    cortexa::architecture::nop ();
+    aarch64::architecture::nop ();
   }
 
   inline __attribute__ ((always_inline)) void
   brk (void)
   {
-    cortexa::architecture::bkpt ();
+    aarch64::architecture::bkpt ();
   }
 
   inline __attribute__ ((always_inline)) void
   wfi (void)
   {
-    cortexa::architecture::wfi ();
+    aarch64::architecture::wfi ();
   }
 
   // --------------------------------------------------------------------------
@@ -169,6 +156,6 @@ namespace micro_os_plus::architecture
 
 // ----------------------------------------------------------------------------
 
-#endif // MICRO_OS_PLUS_ARCHITECTURE_CORTEXA_INSTRUCTIONS_INLINES_H_
+#endif // MICRO_OS_PLUS_ARCHITECTURE_AARCH64_INSTRUCTIONS_INLINES_H_
 
 // ----------------------------------------------------------------------------