+++ include/micro-os-plus/architecture-aarch64/instructions.h
@@ -1,7 +1,7 @@
 /*
  * This file is part of the µOS++ distribution.
  *   (https://github.com/micro-os-plus/)
- * Copyright (c) 2022 Liviu Ionescu.
+ * Copyright (c) 2017 Liviu Ionescu.
  *
  * Permission to use, copy, modify, and/or distribute this software
  * for any purpose is hereby granted, under the terms of the MIT license.
@@ -10,17 +10,17 @@
  * be obtained from https://opensource.org/licenses/MIT/.
  */
 
-#ifndef MICRO_OS_PLUS_ARCHITECTURE_CORTEXA_INSTRUCTIONS_H_
-#define MICRO_OS_PLUS_ARCHITECTURE_CORTEXA_INSTRUCTIONS_H_
+#ifndef MICRO_OS_PLUS_ARCHITECTURE_AARCH64_INSTRUCTIONS_H_
+#define MICRO_OS_PLUS_ARCHITECTURE_AARCH64_INSTRUCTIONS_H_
 
 // ----------------------------------------------------------------------------
 
-#include <micro-os-plus/architecture-cortexa/defines.h>
+#include <micro-os-plus/architecture-aarch64/defines.h>
 
 #include <stdint.h>
 
 // ----------------------------------------------------------------------------
-// Declarations of Cortex-A functions to wrap architecture instructions.
+// Declarations of AArch64 functions to wrap architecture instructions.
 
 #if defined(__cplusplus)
 extern "C"
@@ -28,26 +28,46 @@ extern "C"
 #endif // defined(__cplusplus)
 
   // --------------------------------------------------------------------------
-  // Architecture registers getters and mutators in C.
+  // Architecture assembly instructions in C.
 
   /**
-   * Main Stack Pointer getter.
+   * `nop` instruction.
    */
-  static cortexa_architecture_register_t
-  cortexa_architecture_get_msp (void);
+  static void
+  aarch64_architecture_nop (void);
 
-  // TODO: add setter.
+  /**
+   * `bkpt` instruction.
+   */
+  static void
+  aarch64_architecture_bkpt (void);
+
+  /**
+   * `wfi` instruction.
+   */
+  static void
+  aarch64_architecture_wfi (void);
 
   // --------------------------------------------------------------------------
   // Portable architecture assembly instructions in C.
 
   /**
-   * Stack Pointer getter.
+   * `nop` instruction.
    */
-  static micro_os_plus_architecture_register_t
-  micro_os_plus_architecture_get_sp (void);
+  static void
+  micro_os_plus_architecture_nop (void);
 
-  // TODO: add setter.
+  /**
+   * `break` instruction.
+   */
+  static void
+  micro_os_plus_architecture_brk (void);
+
+  /**
+   * `wfi` instruction.
+   */
+  static void
+  micro_os_plus_architecture_wfi (void);
 
   // --------------------------------------------------------------------------
 
@@ -59,42 +79,62 @@ extern "C"
 
 #if defined(__cplusplus)
 
-namespace cortexa::architecture::registers
+namespace aarch64::architecture
 {
   // --------------------------------------------------------------------------
-  // Architecture getters in C++.
+  // Architecture assembly instructions in C++.
 
   /**
-   * Main Stack Pointer getter.
+   * The assembler `nop` instruction.
    */
-  register_t
-  msp (void);
+  void
+  nop (void);
 
-  // TODO: add setter.
+  /**
+   * The assembler `bkpt` instruction.
+   */
+  void
+  bkpt (void);
+
+  /**
+   * The assembler `wfi` instruction.
+   */
+  void
+  wfi (void);
 
   // --------------------------------------------------------------------------
-} // namespace cortexa::architecture::registers
+} // namespace aarch64::architecture
 
-namespace micro_os_plus::architecture::registers
+namespace micro_os_plus::architecture
 {
   // --------------------------------------------------------------------------
   // Portable architecture assembly instructions in C++.
 
   /**
-   * Main Stack Pointer getter.
+   * The assembler `nop` instruction.
    */
-  register_t
-  sp (void);
+  void
+  nop (void);
 
-  // TODO: add setter.
-  
+  /**
+   * The assembler `bkpt` instruction.
+   */
+  void
+  brk (void);
+
+  /**
+   * The assembler `wfi` instruction.
+   */
+  void
+  wfi (void);
+
   // --------------------------------------------------------------------------
-} // namespace micro_os_plus::architecture::registers
+} // namespace micro_os_plus::architecture
 
 #endif // defined(__cplusplus)
 
 // ----------------------------------------------------------------------------
 
-#endif // MICRO_OS_PLUS_ARCHITECTURE_CORTEXA_INSTRUCTIONS_H_
+#endif // MICRO_OS_PLUS_ARCHITECTURE_AARCH64_INSTRUCTIONS_H_
 
 // ----------------------------------------------------------------------------